refactor: move directory location
